refactoring main js create temlates for arrays of answers create template for logo section
