+++ js/main.js
@@ -1,7 +1,5 @@
-import welcomeTemplate from './templates/start/welcomeTemplate';
 import renderScreen from './utils/renderScreen';
-import {initialState} from './data/data';
 
-renderScreen(welcomeTemplate, initialState);
+renderScreen();
