+++ js/templates/parts/logoTemplate.js
@@ -1,0 +1,5 @@
+const logoTemplate = () => {
+  return `<section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>`;
+};
+
+export default logoTemplate;
