+++ js/templates/result/failResultTemplate.js
@@ -1,7 +1,8 @@
+import logoTemplate from '../parts/logoTemplate';
+
 const failResultTemplate = (level) => {
   return `<section class="main main--result">
-  <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
-
+  ${logoTemplate()}
   <h2 class="title">Вы проиграли</h2>
   <div class="main-stat">Ничего, вам повезет в следующий раз</div>
   <span role="button" tabindex="0" class="main-replay">Сыграть ещё раз</span>
