+++ js/templates/result/successResultTemplate.js
@@ -1,7 +1,8 @@
+import logoTemplate from '../parts/logoTemplate';
+
 const successResultTemplate = (level) => {
   return `<section class="main main--result">
-  <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
-
+  ${logoTemplate()}
   <h2 class="title">Вы настоящий меломан!</h2>
   <div class="main-stat">За&nbsp;2&nbsp;минуты<br>вы&nbsp;отгадали 4&nbsp;мелодии</div>
   <span class="main-comparison">Это&nbsp;лучше чем у&nbsp;80%&nbsp;игроков</span>
