+++ js/templates/start/welcomeTemplate.js
@@ -1,5 +1,7 @@
+import logoTemplate from '../parts/logoTemplate';
+
 const welcomeTemplate = (level) => `<section class="main main--welcome">
-  <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
+  ${logoTemplate()}
   <button class="main-play">Начать игру</button>
   <h2 class="title main-title">Правила игры</h2>
   <p class="text main-text">
